Add createCollection function to CollectionBuffers
Necessary to create an empty collection from the buffers. Not yet sure
if this is the best way to go, several possible problems at the moment:
- Collections are created during reading (e.g. for ROOT), since they own
the buffers that the collection buffers keep as pointers internally, the
buffers are probably usesless at the point where the collection would be
used in the Frame/Event.
- Ideally there is a way for the collections to be created from the
buffers (e.g. via a constructor). That would require a way to get only
the buffers, and additionally also make the buffers the owners of the
data (and the CollectionData owner of the buffers?)