Make collections constructible from buffers
Also introduce the possibility to create thes buffers almost out of thin
air. This needs a refactor, where the use of the original collection to
create the buffers can be removed. Currently we use this mechanism
because this makes it possible to use the existing ROOT metadata
handling and is in a sense backwards compatible with existing files.
Ideally we can create the buffers from metadata directly without having
to go through creating a collection first and then "stealing" its
buffers.